change(web): CustomizationCoord -> InputEventCoord
